[refactor]: rename exception module class file names snake_case
