+++ spatialetl/exception/__init__.py
@@ -9,8 +9,8 @@ EXCEPTION (:mod:`builder.exception`)
 """
 from __future__ import division, print_function, absolute_import
 
-from .CoverageError import CoverageError
-from .VariableNameError import VariableNameError
+from .coverage_error import CoverageError
+from .variable_name_error import VariableNameError
 
 __all__ = ["CoverageError",
            "VariableNameError",
