+++ spatialetl/exception/coverage_error.py
