+++ spatialetl/exception/not_found_in_rank_error.py
